+++ CHANGELOG.md
@@ -8,9 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `!promo` now uses hangout detection determined by the `irc/idle-detection/threshold` setting.
+- `!afterstream` command without argument now prints a help message ([#26]).
 
 ### Changed
 - Fixed bug where `Settings` frontend would make the value into a string before sending it to backend.
+
+[#26]: https://github.com/udoprog/setmod/issues/26
 
 [Unreleased]: https://github.com/udoprog/setmod/compare/0.2.4...HEAD